Fix: Deleting CronJon function on app.ts because vercel not picking up Deleting CronJob.ts because not used anymore Moving CronJob controller from user controller into its own controller (cronjob-controller.ts) Deleting unused multer-middleware becuase currently using multer-cloudinary.ts
