+++ src/routers/user-router.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express';
 import {
   addMember,
-  CronJob,
   deleteMember,
   editMember,
   extendMember,
@@ -19,8 +18,8 @@ import {
 } from '../controllers/user-controller';
 
 import { authUser } from '../middlewares/auth-middleware';
-import { uploadProfilePicture, uploads } from '../middlewares/multer-middleware';
 import { upload } from '../utils/multer-cloudinary';
+import { CronJob } from '../controllers/cronJob-controller';
 
 export const routerUser = Router();
 export const routerDaily = Router();
@@ -30,13 +29,13 @@ routerUser.post('/loginMember', upload.none(), loginMember);
 routerUser.get('/getMember/:id', getMember);
 routerUser.get('/getAllMember', authUser, getAllMember);
 routerUser.patch('/editMember/:id', authUser, upload.single('image'), editMember);
-routerUser.post('/visit/:id', recordVisit);
+routerUser.post('/visit/:id', authUser, recordVisit);
 routerUser.get('/getTodayVisit', authUser, getTodayVisit);
 routerUser.get('/getActiveCount', authUser, getCountMemberActive);
 routerUser.get('/getVisitLog', authUser, getLogVisit);
 routerUser.post('/getpayment', authUser, getAllPayment);
 routerUser.post('/extendMember', authUser, upload.none(), extendMember);
-routerUser.get('/getnotif', getAllNotifications);
+routerUser.get('/getnotif', authUser, getAllNotifications);
 routerUser.delete('/deleteMember', authUser, deleteMember);
 routerUser.get('/getExpiredMember', authUser, getExpiredMember);
 routerUser.post('/visitlog', authUser, getVisitLog);
